Move conversion functions to astria-eyre
