+++ crates/astria-sequencer/src/assets/state_ext.rs
@@ -1,9 +1,12 @@
 use astria_core::primitive::v1::asset;
-use astria_eyre::eyre::{
-    bail,
-    OptionExt as _,
-    Result,
-    WrapErr as _,
+use astria_eyre::{
+    anyhow_to_eyre,
+    eyre::{
+        bail,
+        OptionExt as _,
+        Result,
+        WrapErr as _,
+    },
 };
 use async_trait::async_trait;
 use borsh::{
@@ -20,8 +23,6 @@ use tendermint::abci::{
     EventAttributeIndexExt as _,
 };
 use tracing::instrument;
-
-use crate::utils::anyhow_to_eyre;
 
 /// Newtype wrapper to read and write a denomination trace from rocksdb.
 #[derive(BorshSerialize, BorshDeserialize, Debug)]
