+++ crates/astria-sequencer/src/service/info/mod.rs
@@ -32,12 +32,12 @@ use tracing::{
 
 mod abci_query_router;
 
-use astria_eyre::eyre::Result;
-
-use crate::{
-    state_ext::StateReadExt,
-    utils::anyhow_to_eyre,
+use astria_eyre::{
+    anyhow_to_eyre,
+    eyre::Result,
 };
+
+use crate::state_ext::StateReadExt;
 
 #[derive(Clone)]
 pub(crate) struct Info {
