+++ crates/astria-sequencer/src/utils.rs
@@ -4,7 +4,6 @@ use astria_core::{
     Protobuf as _,
 };
 use astria_eyre::eyre::{
-    eyre,
     Result,
     WrapErr as _,
 };
@@ -18,16 +17,6 @@ impl<'a> std::fmt::Display for Hex<'a> {
         }
         Ok(())
     }
-}
-
-pub(crate) fn anyhow_to_eyre(anyhow_error: anyhow::Error) -> astria_eyre::eyre::Report {
-    let boxed: Box<dyn std::error::Error + Send + Sync> = anyhow_error.into();
-    eyre!(boxed)
-}
-
-pub(crate) fn eyre_to_anyhow(eyre_error: astria_eyre::eyre::Report) -> anyhow::Error {
-    let boxed: Box<dyn std::error::Error + Send + Sync> = eyre_error.into();
-    anyhow::anyhow!(boxed)
 }
 
 pub(crate) fn cometbft_to_sequencer_validator(
@@ -94,34 +83,6 @@ mod pubkey {
         };
         AstriaKey {
             sum,
-        }
-    }
-}
-
-mod test {
-    #[test]
-    fn anyhow_to_eyre_preserves_source_chain() {
-        let mut errs = ["foo", "bar", "baz", "qux"];
-        let anyhow_error = anyhow::anyhow!(errs[0]).context(errs[1]).context(errs[2]);
-        let eyre_from_anyhow = super::anyhow_to_eyre(anyhow_error).wrap_err(errs[3]);
-
-        errs.reverse();
-        for (i, err) in eyre_from_anyhow.chain().enumerate() {
-            assert_eq!(errs[i], &err.to_string());
-        }
-    }
-
-    #[test]
-    fn eyre_to_anyhow_preserves_source_chain() {
-        let mut errs = ["foo", "bar", "baz", "qux"];
-        let eyre_error = astria_eyre::eyre::eyre!(errs[0])
-            .wrap_err(errs[1])
-            .wrap_err(errs[2]);
-        let anyhow_from_eyre = super::eyre_to_anyhow(eyre_error).context(errs[3]);
-
-        errs.reverse();
-        for (i, err) in anyhow_from_eyre.chain().enumerate() {
-            assert_eq!(errs[i], &err.to_string());
         }
     }
 }
